Split client and usage
